Added visual feedback for correct/incorrect
An O or X icon now appears over the target word based on correctness. There is currently no sound played along with it.

Incorrect display is currently commented out due to no way to detect if a word spoken is incorrect (Line 325).

Also cleaned up a few resources that weren't being disposed of